pyproject.toml
[tool.poetry]
name = "billing"
version = "0.1.0"
description = ""
authors = ["Danil <dk@forcode.pro>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
Django = "^4.1"
psycopg2 = "^2.9.3"
env = "^0.1.0"
poetry-dotenv-plugin = "^0.1.0-alpha.2"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
